move replace_values_in_json method to json_utils
