+++ src/lib/core/grafana_config_generator.py
@@ -2,35 +2,10 @@ import os
 import json
 import logging
 
+from lib.core import json_utils as ju
+
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
-
-
-def replace_values_in_json(json_data: json, replacements: dict) -> dict:
-    """
-    Replaces values in a JSON object based on a dictionary of replacements.
-
-    Args:
-        json_data (dict): The JSON object.
-        replacements (dict): Dictionary with keys as the values to be replaced and
-                            corresponding values as the replacements.
-
-    Returns:
-        dict: Updated JSON object.
-    """
-
-    def replace_recursive(obj):
-        if isinstance(obj, dict):
-            for key, value in obj.items():
-                obj[key] = replace_recursive(value)
-        elif isinstance(obj, list):
-            for i, item in enumerate(obj):
-                obj[i] = replace_recursive(item)
-        elif isinstance(obj, str) and obj in replacements:
-            obj = replacements[obj]
-        return obj
-
-    return replace_recursive(json_data)
 
 
 def generate_cloudwatch_dashboard_model(
@@ -48,7 +23,7 @@ def generate_cloudwatch_dashboard_model(
         dashboard_data (dict): Dashboard json model.
     """
     dashboard_path = os.path.join(
-        "infra_tooling_account", "grafana", "cloudwatch_dashboard.sample.json"
+        "infra_tooling_account", "grafana", "cloudwatch_dashboard.template.json"
     )
     with open(dashboard_path) as json_file:
         json_data = json.load(json_file)
@@ -58,7 +33,7 @@ def generate_cloudwatch_dashboard_model(
         "<<LOG_GROUP_ARN>>": cloudwatch_log_group_arn,
         "<<ACCOUNT_ID>>": account_id,
     }
-    dashboard_data = replace_values_in_json(json_data, replacements)
+    dashboard_data = ju.replace_values_in_json(json_data, replacements)
 
     return dashboard_data
 
@@ -77,7 +52,7 @@ def generate_timestream_dashboard_model(
         dashboard_data (dict): Dashboard json model.
     """
     dashboard_path = os.path.join(
-        "infra_tooling_account", "grafana", f"{resource_type}_dashboard.sample.json"
+        "infra_tooling_account", "grafana", f"{resource_type}_dashboard.template.json"
     )
     try:
         with open(dashboard_path) as json_file:
@@ -90,7 +65,7 @@ def generate_timestream_dashboard_model(
         "<<DATABASE_NAME>>": f'"{timestream_database_name}"',
         "<<DATABASE_TABLE>>": f'"{timestream_table_name}"',
     }
-    dashboard_data = replace_values_in_json(json_data, replacements)
+    dashboard_data = ju.replace_values_in_json(json_data, replacements)
 
     return dashboard_data
 
@@ -145,7 +120,7 @@ def generate_dashboards_config(resource_types: list) -> dict:
     dashboards_sections = []
     for resource_type in resource_types:
         dashboard_path = os.path.join(
-        "infra_tooling_account", "grafana", f"{resource_type}_dashboard.sample.json"
+        "infra_tooling_account", "grafana", f"{resource_type}_dashboard.template.json"
         )
         if os.path.exists(dashboard_path):
             dashboard_section = {
@@ -185,7 +160,7 @@ def generate_user_data_script(
         user_data_content (str): User data with set of commands.
     """
     user_data_file_path = os.path.join(
-        "infra_tooling_account", "grafana", "grafana_user_data.sample.sh"
+        "infra_tooling_account", "grafana", "grafana_user_data.template.sh"
     )
     with open(user_data_file_path, "r") as user_data_file:
         user_data_content = user_data_file.read()
